Added search option to admin panel users.
Reloads page and posts a hidden form that contains the new search value.
A page reload was needed in order to re-initialize the materialize
pagination addon. (new total page numers from new search query)

However, this does not change the dynamic user loading when pressing the
page numbers. The maximum users loaded per ajax request is (at the time
of writing set to a LIMIT of 10).